refactor: start refactor app architecture
