--- utils/cli/env.js
@@ -1,8 +1,0 @@
-const parseEnv = () => {
-  const envObj = process.env;
-  Object.keys(envObj)
-    .filter((item) => item.startsWith("RSS_"))
-    .forEach((key) => console.log(`${key}=${envObj[key]};`));
-};
-
-parseEnv();
