--- utils/zip/files/fileToCompress.txt
@@ -1,1 +1,0 @@
-Compress me!
